Version 0.0.1 of clapp (#1)
Clapp is a golang package designed to make it quicker and easier to
bootstrap an app that runs from the command line.

It provides a slightly more convenient (opionated) way to define
all the commands in the app and to load configuration.

It allows configuration to specificed at 3 levels: file, env, and
cli flags, in that order of precedence. The configuration values
can be loaded into a struct which will be stored within the context
making it easy to access the config in any area of the app where
the context is passed.

It also provides a logger (zerolog) and a filesystem (afero), which
are very common dependencies for cli tools.