Switch generic interfaces to cobra.Command (#3)
In a couple of places the library was designed to be unneccessarily
generic. The command handler functions would expect a generic
interface which would need to be type asserted before you can do
anything useful with it. In these places I replaced it for the
concrete implementation of cobra.Command, to make the library
simpler to use.